pyproject.toml
[project]
name = "projet7"
version = "0.1.0"
description = ""
authors = [
    {name = "Apinsun", email = "aranudpinsun@gmail.com"}
]
readme = "README.md"
# On cible ta version (3.12) et on reste sous 3.15 pour satisfaire FAISS
requires-python = ">=3.12,<3.15"

dependencies = [
    "pandas (>=3.0.1,<4.0.0)",
    "requests (>=2.33.0,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "mistralai (>=2.1.3,<3.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "langchain (>=1.2.13,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-huggingface (>=1.2.1,<2.0.0)",
    "langchain-mistralai (>=1.1.2,<2.0.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "ragas (>=0.4.3,<0.5.0)",
    "datasets (>=4.8.4,<5.0.0)",
    "instructor (<1.3.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
